+++ src/main/java/net/helix/hlx/conf/ConfigFactory.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;