+++ src/main/java/net/helix/hlx/crypto/Merkle.java
@@ -3,12 +3,7 @@ package net.helix.hlx.crypto;
 import net.helix.hlx.model.Hash;
 import org.bouncycastle.util.encoders.Hex;
 
-import java.io.BufferedWriter;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.File;
+import java.io.*;
 import java.util.Arrays;
 
 public class Merkle {