+++ src/main/java/net/helix/hlx/network/Node.java
@@ -11,12 +11,10 @@ import net.helix.hlx.crypto.SpongeFactory;
 import net.helix.hlx.model.Hash;
 import net.helix.hlx.model.HashFactory;
 import net.helix.hlx.model.TransactionHash;
-import net.helix.hlx.model.persistables.Bundle;
+import net.helix.hlx.service.Graphstream;
 import net.helix.hlx.service.milestone.LatestMilestoneTracker;
 import net.helix.hlx.service.snapshot.SnapshotProvider;
 import net.helix.hlx.storage.Tangle;
-import net.helix.hlx.service.Graphstream;
-
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
@@ -180,7 +178,7 @@ public class Node {
      * traffic - so a balance is sought between speed and resource utilization.
      *
      */
-    private Runnable spawnNeighborDNSRefresherThread() {
+    Runnable spawnNeighborDNSRefresherThread() {
         return () -> {
             if (configuration.isDnsResolutionEnabled()) {
                 log.info("Spawning Neighbor DNS Refresher Thread");