+++ src/main/java/net/helix/hlx/service/milestone/MSS.java
@@ -6,10 +6,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-
-import java.util.concurrent.Executors;
 
 public class MSS {