+++ src/main/java/net/helix/hlx/service/snapshot/impl/SnapshotServiceImpl.java
@@ -19,6 +19,8 @@ import net.helix.hlx.utils.dag.DAGHelper;
 import net.helix.hlx.utils.dag.TraversalException;
 import net.helix.hlx.utils.log.ProgressLogger;
 import net.helix.hlx.utils.log.interval.IntervalProgressLogger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -26,9 +28,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Creates a service instance that allows us to access the business logic for {@link Snapshot}s.<br />