+++ src/test/java/net/helix/hlx/benchmarks/BenchmarkRunner.java
@@ -1,0 +1,46 @@
+package net.helix.hlx.benchmarks;
+
+import net.helix.hlx.benchmarks.dbbenchmark.RocksDbBenchmark;
+import org.junit.Test;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.runner.Runner;
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.Options;
+import org.openjdk.jmh.runner.options.OptionsBuilder;
+
+import java.util.concurrent.TimeUnit;
+
+public class BenchmarkRunner {
+
+    @Test
+    public void launchDbBenchmarks() throws RunnerException {
+        Options opts = new OptionsBuilder()
+                .include(RocksDbBenchmark.class.getName() + ".*")
+                .mode(Mode.Throughput) //Mode.AverageTime
+                .timeUnit(TimeUnit.MILLISECONDS)
+                .warmupIterations(1) //5
+                .forks(1)
+                .measurementIterations(1) //10
+                .shouldFailOnError(true)
+                .shouldDoGC(false)
+                .build();
+
+        //possible to do assertions over run results
+        new Runner(opts).run();
+    }
+
+    @Test
+    public void launchCryptoBenchmark() throws RunnerException {
+        Options opts = new OptionsBuilder()
+                .include(this.getClass().getPackage().getName() + ".crypto")
+                .mode(Mode.Throughput)
+                .timeUnit(TimeUnit.SECONDS)
+                .warmupIterations(5)
+                .forks(1)
+                .measurementIterations(10)
+                .shouldFailOnError(true)
+                .shouldDoGC(false)
+                .build();
+        new Runner(opts).run();
+    }
+}